+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 * Recreate initial alembic migrations. (rudaporto)
 * Update all tests and test data to fit the changes in the models. (rudaporto)
 * LEICA-29: Add initial custom route factory for each model except JobLocation. (rudaporto)
+* LEICA-38: Add uploaded_by to Asset (ericof).
 * BODY-31: fixed briefy.ws issue. (rudaporto)
 * LEICA-30: return comments list on the result payload of Jobs and Assets. (rudaporto)
 * LEICA-31: Run asset.update_metada() method every time afeter asset model instance change. (rudaporto)